// src/app/home/models/tb-supplier.ts
export class TbSupplier {
  supSerial?: number = null;
  supAddr1?: string = null;
  supAddr1Th?: string = null;
  supAmphur?: string = null;
  supAmphurTh?: string = null;
  supCode?: string = null;
  supFax?: string = null;
  supMobile?: string = null;
  supNameEng?: string = null;
  supNameTh?: string = null;
  supPhone?: string = null;
  supPostcode?: string = null;
  supPplno?: string = null;
  supProvinceId?: number = null;
  supTumbon?: string = null;
  supTumbonTh?: string = null;
  supauthpplType?: number = null;
  supbranchCode?: string = null;
  supbranchName?: string = null;
  supcargoPort?: number = null;
  supdeliveryTag?: number = null;
  suplayoutVers?: number = null;
  supplantCode?: string = null;
  supprodType?: number = null;
  supregisBoi19?: string = null;
  supregisBoippl?: string = null;
  supregisN019?: string = null;
  supreleasePort?: number = null;
  supstsFlag?: string = null;
  supstsFlagDesc?: string = null;
  suptaxId?: string = null;
  suptaxIncentive?: string = null;
  supuseFlag?: number = null;
  supuseType?: number = null;

  provinceDesc?: string = null;

  offset?: number = null;
  limit?: number = null;
  createBy?: string = null;
  updateBy?: string = null;
}